Add theory library and integrate tests
Begin import of coltrane theory library, this is needed because the
current gem has a lot of features I don't need and is incomplatible with
the latest version of Rails. Mostly this all just works but there is a
few failed tests because of an issue with the defining of methods.

Theory library is from the MIT licensed coltrane gem found here
https://github.com/pedrozath/coltrane